requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=5.2.2
python-dotenv>=1.0.1
PyYAML>=6.0.2
jmespath>=1.0.1
sentence-transformers>=3.0.1
sentencepiece>=0.2.1
protobuf>=3.20.0
Pillow>=10.4.0
playwright>=1.40.0
supabase>=2.0.0
transformers>=4.35.0
torch>=2.0.0
torchvision>=0.15.0
